separate config header in minnesota
